requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
gunicorn==21.2.0
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.2
packaging==23.2
python-dotenv==1.0.1
Werkzeug==3.0.1
zipp==3.13.0

WTForms~=3.1.2
alembic~=1.13.1
SQLAlchemy~=2.0.29